Started basic GA implementation
Started filling out the basics on the genetic algorithm stuff.  Have
only really started with the basic player structure.  Will probably
change it is things start to develop.  Haven't even started the
management structure for building and running the genetic algorithms.
Just occurred to me a UI for running the GA would be nice.